use base-resources for all firestore operations
