+++ src/app/core/core.model.ts
@@ -5,3 +5,13 @@ export interface IdLike {
 export interface MaybeIdLike {
   id?: string;
 }
+
+export class PhenonetUser {
+  nickname: string;
+  firstname: string;
+  lastname: string;
+  locale: string;
+
+  following_users?: string[];
+  following_individuals?: string[];
+}
