--- src/app/shared/models/user.model.ts
@@ -1,9 +1,0 @@
-export class PhenonetUser {
-  nickname: string;
-  firstname: string;
-  lastname: string;
-  locale: string;
-
-  following_users?: string[];
-  following_individuals?: string[];
-}
